--- apps/sqltools/src/components/TutorialAlert.vue
@@ -1,16 +1,0 @@
-<template>
-  <div class="alert">
-    <div>SQLTools is a SQL table builder for Postgres, MySQL, SQL Server, SQLite, and Redshift databases from the makers of  <a class="text-primary" href="https://beekeeperstudio.io" target="_blank">Beekeeper Studio</a>. We also provide some simple sql table starter templates, check them out above 👆.</div>
-  </div>
-
-</template>
-<script lang="ts">
-import Vue from 'vue'
-export default Vue.extend({
-  data() {
-    return {
-      content: ""
-    }
-  }
-})
-</script>
